MINOR: Make TopicPartitionBookkeeper and TopicPartitionEntry top level
This is the first step towards refactoring the `TransactionManager` so
that it's easier to understand and test. The high level idea is to push
down behavior to `TopicPartitionEntry` and `TopicPartitionBookkeeper`
and to encapsulate the state so that the mutations can only be done via
the appropriate methods.

Inner classes have no mechanism to limit access from the outer class,
which presents a challenge when mutability is widespread, like it is the
case here.

As a first step, we make `TopicPartitionBookkeeper` and
`TopicPartitionEntry` top level and rename them to make their intended
usage clear. To make the review easier, we don't change anything else
except access changes required for the code to compile.

The next PR will contain the rest of the refactoring.
